magic追加、LICENSE.txt追加

+++ src/main/java/com/hutuneko/magic_chants/api/chat/ChantScorer.java
@@ -5,47 +5,47 @@ import java.util.stream.Collectors;
 
 public class ChantScorer {
 
-    public static double score(String chant) {
-        if (chant == null || chant.isBlank()) return 0;
+    public static float score(String chant) {
+        if (chant == null || chant.isBlank()) return 0f;
 
         // 全角スペース・句読点・記号を区切りに正規化
         String normalized = chant.replaceAll("[、。・.,;!?！？　]", " ").trim();
         String[] words = normalized.split("\\s+");
-        if (words.length == 0) return 0;
+        if (words.length == 0) return 0f;
 
         // ① 繰り返し率（多いほどリズム感）
-        double repScore = repetitionScore(words);
+        float repScore = (float) repetitionScore(words);
 
         // ② 語長の安定度（標準偏差が低いほど高得点）
-        double rhythmScore = rhythmScore(words);
+        float rhythmScore = (float) rhythmScore(words);
 
         // ③ 音の繋がり（語尾と次語の頭が一致する率）
-        double flowScore = flowScore(words);
+        float flowScore = (float) flowScore(words);
 
         // ④ 文字多様性（ユニーク文字 / 全文字）
-        double diversityScore = diversityScore(normalized);
+        float diversityScore = (float) diversityScore(normalized);
 
         // ⑤ 架空語率（漢字・カタカナ・記号など）
-        double fictionScore = fictionScore(normalized);
+        float fictionScore = (float) fictionScore(normalized);
 
         // 総合スコア（重み平均）
-        double finalScore = 10 * (
-                0.25 * repScore +
-                        0.25 * rhythmScore +
-                        0.2  * flowScore +
-                        0.2  * diversityScore +
-                        0.1  * fictionScore
+        float finalScore = 10f * (
+                0.25f * repScore +
+                        0.25f * rhythmScore +
+                        0.2f  * flowScore +
+                        0.2f  * diversityScore +
+                        0.1f  * fictionScore
         );
 
-        return Math.round(finalScore * 10.0) / 10.0; // 小数1桁
+        return Math.round(finalScore * 10f) / 10f; // 小数1桁
     }
 
     private static double repetitionScore(String[] words) {
         Map<String, Long> freq = Arrays.stream(words)
                 .collect(Collectors.groupingBy(w -> w, Collectors.counting()));
         long max = freq.values().stream().max(Long::compare).orElse(1L);
-        double ratio = (double) max / words.length; // 最大出現語の比率
-        return Math.max(0, Math.min(1, ratio)); // 0〜1
+        double ratio = (double) max / words.length;
+        return Math.max(0, Math.min(1, ratio));
     }
 
     private static double rhythmScore(String[] words) {
@@ -54,7 +54,6 @@ public class ChantScorer {
         double variance = Arrays.stream(lens)
                 .map(l -> Math.pow(l - mean, 2)).sum() / lens.length;
         double stddev = Math.sqrt(variance);
-        // 安定しているほど高得点（ばらつきが少ない）
         return 1.0 / (1.0 + stddev / mean);
     }
